Break down public/components/utils/status-box.tsx into separate files
