+++ frontend/public/components/utils/status-box/LoadingInline.tsx
@@ -1,0 +1,5 @@
+import * as React from 'react';
+import { Loading } from '.';
+
+export const LoadingInline: React.FC<{}> = () => <Loading className="co-m-loader--inline" />;
+LoadingInline.displayName = 'LoadingInline';
